% Auto-generated: do not edit by hand
\name{downloadButton}

\alias{downloadButton}

\title{DownloadButton component}

\description{

}

\usage{
downloadButton(children=NULL, id=NULL, className=NULL, hidden=NULL, target=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. The children of this component}

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. Often used with CSS to style elements with common properties.}

\item{hidden}{Character. Prevents rendering of given element, while keeping child elements, e.g. script elements, active.}

\item{target}{Character. Defines a default value which will be displayed in the element on page load.}
}
